requirements.txt
colorama==0.4.6
requests==2.32.5
pillow==11.3.0
qrcode==8.2